Move interaction code with aws services under out of actuator
