+++ pkg/cloud/aws/actuators/machine/actuator_test.go
@@ -3,7 +3,6 @@ package machine
 import (
 	"errors"
 	"fmt"
-	"reflect"
 	"strings"
 	"testing"
 	"time"
@@ -343,82 +342,6 @@ func mockRegisterInstancesWithLoadBalancer(mockAWSClient *mockaws.MockClient, cr
 	// spec for create and for update (3 * 2 = 6)
 	for i := 0; i < 6; i++ {
 		mockAWSClient.EXPECT().RegisterInstancesWithLoadBalancer(gomock.Any())
-	}
-}
-
-func TestRemoveDuplicatedTags(t *testing.T) {
-	cases := []struct {
-		tagList  []*ec2.Tag
-		expected []*ec2.Tag
-	}{
-		{
-			// empty tags
-			tagList:  []*ec2.Tag{},
-			expected: []*ec2.Tag{},
-		},
-		{
-			// no duplicate tags
-			tagList: []*ec2.Tag{
-				{Key: aws.String("clusterID"), Value: aws.String("test-ClusterIDValue")},
-			},
-			expected: []*ec2.Tag{
-				{Key: aws.String("clusterID"), Value: aws.String("test-ClusterIDValue")},
-			},
-		},
-		{
-			// multiple duplicate tags
-			tagList: []*ec2.Tag{
-				{Key: aws.String("clusterID"), Value: aws.String("test-ClusterIDValue")},
-				{Key: aws.String("clusterSize"), Value: aws.String("test-ClusterSizeValue")},
-				{Key: aws.String("clusterSize"), Value: aws.String("test-ClusterSizeDuplicatedValue")},
-			},
-			expected: []*ec2.Tag{
-				{Key: aws.String("clusterID"), Value: aws.String("test-ClusterIDValue")},
-				{Key: aws.String("clusterSize"), Value: aws.String("test-ClusterSizeValue")},
-			},
-		},
-	}
-
-	for i, c := range cases {
-		actual := removeDuplicatedTags(c.tagList)
-		if !reflect.DeepEqual(c.expected, actual) {
-			t.Errorf("test #%d: expected %+v, got %+v", i, c.expected, actual)
-		}
-	}
-}
-
-func TestBuildEC2Filters(t *testing.T) {
-	filter1 := "filter1"
-	filter2 := "filter2"
-	value1 := "A"
-	value2 := "B"
-	value3 := "C"
-
-	inputFilters := []providerconfigv1.Filter{
-		{
-			Name:   filter1,
-			Values: []string{value1, value2},
-		},
-		{
-			Name:   filter2,
-			Values: []string{value3},
-		},
-	}
-
-	expected := []*ec2.Filter{
-		{
-			Name:   &filter1,
-			Values: []*string{&value1, &value2},
-		},
-		{
-			Name:   &filter2,
-			Values: []*string{&value3},
-		},
-	}
-
-	got := buildEC2Filters(inputFilters)
-	if !reflect.DeepEqual(expected, got) {
-		t.Errorf("failed to buildEC2Filters. Expected: %+v, got: %+v", expected, got)
 	}
 }
